backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy==2.0.23
alibabacloud-credentials
alibabacloud_dysmsapi20170525
alibabacloud_tea_openapi
alibabacloud_tea_util
cryptography==41.0.7
matplotlib
numpy<2
numexpr>=2.10.2
bottleneck>=1.4.2
Pillow
paddleocr>=2.8.1
paddlepaddle>=3.0.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
requests==2.31.0
waitress==3.0.0
openai>=1.30.0
jieba==0.42.1
torch==2.9
torchvision==0.24
transformers==5.0.0
accelerate==1.12.0
openai>=1.0
jieba